requirements.txt
discord.py>=2.3.1,<2.4.0
lru-dict>=1.2.0,<1.3.0
pydantic-settings>=2.0.2,<2.1.0
requests>=2.31.0,<2.32.0
